requirements.txt
Flask==1.1.1
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.23.0
python-dateutil==2.2
redis==3.4.1
requests==2.20.1
rq==1.3.0
rq-scheduler==0.9.1
SQLAlchemy==1.3.16
yagmail==0.11.224
